requirements.txt
pyyaml
requests
pypika
mysqlclient
Faker
